Remove drivers and place image dumping code in separate module
